wpl用long long

+++ 08_Extend_OJ/04_Sunnywhy/09/9.8_2/wpl_min_weight_path_length.cpp
@@ -26,7 +26,7 @@ int main(int argc, char *argv[])
         min_heap.push(weight);
     }
 
-    int wpl = 0;
+    long long wpl = 0;
     while (min_heap.size() >= 2)
     {
         long long w1 = min_heap.top();
